+++ src/main/java/com/example/saheel/Service/VeterinaryService.java
@@ -2,7 +2,9 @@ package com.example.saheel.Service;
 
 import com.example.saheel.Api.ApiException;
 import com.example.saheel.Model.Stable;
+import com.example.saheel.Model.StableOwner;
 import com.example.saheel.Model.Veterinary;
+import com.example.saheel.Repository.StableOwnerRepository;
 import com.example.saheel.Repository.StableRepository;
 import com.example.saheel.Repository.VeterinaryRepository;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +16,7 @@ public class VeterinaryService {
 
     private final VeterinaryRepository veterinaryRepository;
 
+    private final StableOwnerRepository stableOwnerRepository;
     private final StableRepository stableRepository;
 
     // get Veterinary by ID - Abeer
@@ -26,18 +29,57 @@ public class VeterinaryService {
         return veterinary;
     }
 
+    public Veterinary searchVeterinaryByName(Integer stableOwner_Id ,String fullName){
+
+        StableOwner stableOwner = stableOwnerRepository.findStableOwnerById(stableOwner_Id);
+        if (stableOwner == null) {
+            throw new ApiException("Error : stable owner is not found");
+        }
+        Veterinary veterinary = veterinaryRepository.findVeterinaryByFullName(fullName);
+        if (veterinary == null){
+            throw new ApiException("Error : veterinary is not found");
+        }
+        return veterinary;
+    }
+
     //add veterinary - Abeer
-    public void addVeterinary(Integer stable_Id , Veterinary veterinary){
+    public void addVeterinary(Integer stableOwner_Id, Integer stable_Id , Veterinary veterinary){
         Stable stable = stableRepository.findStableById(stable_Id);
         if (stable == null){
             throw new ApiException("Error : stable is not fond");
+        }
+        // Check that this stable belongs to the registered owner.
+        if (!stable.getStableOwner().getId().equals(stableOwner_Id)) {
+            throw new ApiException("Unauthorized error: This stable does not belong to the logged-in stable owner");
         }
         veterinary.setStable(stable);
         veterinaryRepository.save(veterinary);
     }
 
+    //move veterinary To Another Stable -Abeer
+    public void moveVeterinaryToAnotherStable(Integer stableOwner_Id, Integer veterinary_Id, Integer stable_Id) {
+
+        Veterinary veterinary = veterinaryRepository.findVeterinaryById(veterinary_Id);
+        if (veterinary == null) {
+            throw new ApiException("Error Trainer is not found");
+        }
+
+        Stable newStable = stableRepository.findStableById(stable_Id);
+        if (newStable == null) {
+            throw new ApiException("Target stable not found");
+        }
+
+        // Check that this stable belongs to the registered owner.
+        if (!newStable.getStableOwner().getId().equals(stableOwner_Id)) {
+            throw new ApiException("Error : You are not the owner of this stable");
+        }
+
+        veterinary.setStable(newStable);
+        veterinaryRepository.save(veterinary);
+    }
+
     //update Veterinary - Abeer
-    public void updateVeterinary(Integer stable_Id ,Integer veterinary_Id, Veterinary veterinary ) {
+    public void updateVeterinary( Integer stableOwner_Id, Integer stable_Id , Integer veterinary_Id, Veterinary veterinary ) {
         Stable stable = stableRepository.findStableById(stable_Id);
         if (stable == null){
             throw new ApiException("Error : stable is not fond");
@@ -47,6 +89,11 @@ public class VeterinaryService {
         if (oldVeterinary == null) {
             throw new ApiException("Error: Veterinary not found");
         }
+        // Check that this stable belongs to the registered owner.
+        if (!stable.getStableOwner().getId().equals(stableOwner_Id)) {
+            throw new ApiException("Error : You are not the owner of this stable");
+        }
+
 
         oldVeterinary.setUsername(veterinary.getUsername());
         oldVeterinary.setFullName(veterinary.getFullName());
@@ -54,15 +101,21 @@ public class VeterinaryService {
         oldVeterinary.setEmail(veterinary.getEmail());
         oldVeterinary.setYearsOfExperience(veterinary.getYearsOfExperience());
         oldVeterinary.setRating(veterinary.getRating());
+
         veterinaryRepository.save(veterinary);
     }
 
     //delete  Veterinary - Abeer
-    public void deleteVeterinary(Integer veterinary_Id) {
+    public void deleteVeterinary(Integer stableOwner_Id, Integer veterinary_Id) {
         Veterinary veterinary = veterinaryRepository.findVeterinaryById(veterinary_Id);
         if (veterinary == null) {
             throw new ApiException("Error: Veterinary not found");
         }
+        Stable stable = veterinary.getStable();
+        if (stable == null || !stable.getStableOwner().getId().equals(stableOwner_Id)) {
+            throw new ApiException("Error: You do not have permission to delete this veterinary");
+        }
+
         veterinaryRepository.delete(veterinary);
     }
